+++ src/listeners.ts
@@ -6,10 +6,7 @@ export function createListeners(editor: TextEditor) {
     const { color } = evt;
     console.log(`addCustomColor:`, color);
   });
-  editor.addEventListener("colors:addCustomColorToView", (evt: any) => {
-    const { color } = evt;
-    console.log(`addCustomColorToView:`, color);
-  });
+
   editor.addEventListener("colors:removeCustomColor", (evt: any) => {
     const { color } = evt;
     console.log(`removeCustomColor:`, color);
@@ -24,13 +21,6 @@ export function createListeners(editor: TextEditor) {
     const { color } = evt;
     console.log(`addCustomBackgroundColor:`, color);
   });
-  editor.addEventListener(
-    "colors:addCustomBackgroundColorToView",
-    (evt: any) => {
-      const { color } = evt;
-      console.log(`addCustomBackgroundColorToView:`, color);
-    },
-  );
 
   //PAGE LINKS
   editor.addEventListener("pageLink:selectedPage", (evt: any) => {