+++ src/plugins/FontBackground/FontBackgroundUI.ts
@@ -78,7 +78,6 @@ export class FontBackgroundUI extends Plugin implements ColorViewer {
           color,
         };
         this.editor.dispatch("colors:addCustomBackgroundColor", evt);
-        this.editor.dispatch("colors:addCustomBackgroundColorToView", evt);
       }
     };
     input?.click();